pill-counting-py-server/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aioresponses==0.7.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.4.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inference-sdk==0.22.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.23.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
six==1.16.0
supervision==0.23.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.12.1
zipp==3.20.2
